fix: classify mvn test as runtime and gate coverage continue-on-error
Closes #5178
Closes #5172
Related to #5186
Tracker remains #5186
